perubahan tipe data nominal pada finance_mutations dari decimal(20,3) menjadi decimal(65,30) dan finance_mutation_schedules dari double menjadi decimal(65,30)

+++ app/Http/Requests/ValidateFinanceMutationSchedule.php
@@ -52,7 +52,7 @@ class ValidateFinanceMutationSchedule extends FormRequest
             'next_mutation_date' => ['required','date','after:today'],
             'name' => ['required','string','max:100'],
             'currency' => ['required','in:usd,cny,idr'],
-            'nominal' => ['required','integer'],
+            'nominal' => ['required','numeric'],
             'mode' => ['required','in:debit,credit,transfer'],
             'from_wallet_id' => ["bail","required_if:mode,transfer","nullable","integer",Rule::exists(Wallet::class,'id')->where(function($query){
                 $query->whereNull('deleted_at');
